LayoutTests/imported/w3c/web-platform-tests/event-timing/only-observe-firstInput-expected.txt

Generate a 'click' event

FAIL Event Timing: check first-input for a PerformanceObserver observing only first-input. assert_equals: expected 1 but got 0
